requirements.txt
hupper==1.10.2
mypy==0.770
mypy-extensions==0.4.3
PasteDeploy==2.1.0
pkg-resources==0.0.0
plaster==1.0
plaster-pastedeploy==0.7
psycopg2-binary==2.8.4
pyramid==1.10.4
PyYAML==5.3.1
SQLAlchemy==1.3.15
translationstring==1.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
venusian==3.0.0
WebOb==1.8.6
zope.deprecation==4.4.0
zope.interface==5.0.2
